Responsibility to check CPU features is callers'
For the singlepass compiler the AVX feature is required, but in practice
this is only true when floating point code is used and NaNs need to be
normalized. In `nearcore` we have an option to disable the CPU feature
checks in order to enable testing of simpler contracts on machines that
do not support the AVX feature. Most prominently this is QEMU with some
plugins.

We still check the features at compile time, so that seems to be good
enough for now.