Use a condition flag for unsigned results
The `Signed` flag works okay, but it will be set when the stack limit is
anywhere `> 0xF000_0000`. Given that one might want to set the limit to
a value such as `0xFFFF_FFFF` to effectively disable the stack checks,
the `Signed` flag would get set and the runtime would still report an
incorrect stack overflow.