Disentangle UniversalArtifact and UniversalEngine
As of this commit, the `UniversalArtifact` is entirely acting as a
representation of the compiled form of the artifact, either in memory or
on disk.

All of the functionality related to loading the `UniversalArtifact` and
preparing it for execution is instead moved to instantiation time.

While this means that for `Artifact`s, which are instantiated multiple
times, the performance will drop, this also allows 0-copy
deserialization of the artifact itself and similar improvements.