Remove code related to signal handling
The primary motivation for doing this is related to the frame_info
registration that was occuring during instantiation. In particular the
expectation was that the information about function extents and similar
would be present throughout the lifetime of the instance.

Now that we moved any such state away from the Artifacts, the next
obvious place to put the relevant data would be into the `Instance`.
However doing so presents a problem – it would make `wasmer_vm` depend
on `wasmer_engine`, producing a cyclic dependency cycle.

Since this frame information is utilized exclusively for signal-based
trap handling, which we do not and cannot use, it seemed to make sense
to me to rip out the functionality altogether, removing one of the more
tricky/sketchy parts of the codebase in the process.

The expectation is that some traps (such as stack overflow or gas
exceeded) would remain invoked either as a result of failing
instrumented checks in wasmer, with us ensuring there's sufficient
resources or similar that the actual host-side limits are not reached.
In other instances when we'd actually hit an error serious enough to
raise a signal, I think we'd rather crash the interpreter entirely and
figure out the sandboxing bug.

Note that this still leaves some code related to `FRAME_INFO` intact.
This code was used to produce and capture stack traces in
`RuntimeError`s. While most of this code is “dead” at the time of this
commit (we never register any `FRAME_INFO`), it seems plausible that we
could recover some or most of that functionality if we wanted it (we
don't currently utilize anything besides the `TrapCode`) in the future.