Add a concept of an Executable
Executable is a deserialized but not loaded compiled code. This allows
us to implement a mechanism by which we copy/create WebAssembly module
data fewer times before it becomes an instance. Of particular interest,
it can help to avoid allocating the tree of maps and vectors that are
used to represent the module when an Artifact has already been
compiled.