Refactor the handling of locals in singlepass
Among other things this commit ensures that the handling of locals at
compile time is linear to the encoding size of the locals, rather than
the number of locals. This means that with this change the size of
WebAssembly input becomes a reasonable estimation of how long processing
it might take, while before adding many locals, even if unused, would
result in huge amount of processing, even if the locals themselves
occupied just a couple bytes of module encoding. This overall shows
0~50% compile-time improvement on select operations depending on the
number of locals and their encoding.

Runtime wise there are no changes whatsoever and the generated machine
code should behave exactly as before.