Set up the trampoline for Static host functions
These are somewhat interesting in that they still require a trampoline
to be called (in the current construction of the runtime), but the
trampoline will typically be set to `None` (because these are generated
by the runtime, and the host has no idea where to find them.)

We tip-toe around the problem by setting the trampoline up during the
linking phase.

In long term I think we should reduce the reliance on – and the need for
– the trampolines across the board. One thing that comes as immediately
removable as a simplification is any support for dynamic function
imports of the `fn(Ctx, &[Value])` form. Similarly, instead of passing
the VMContext as a argument, we could probably set up a reserved
register with a value for this on the host side as well, especially now
that inline assembly is a thing.