Remove the `Executable::load` and `Engine::load`
These methods attempt to make it straightforward to load a module into a
store when the engine contained therein is already a dynamic object.
While there are some instances of it, for all cases where it matters for
us, obtaining a specific type of engine is going to be pretty
straightforward (`UniversalEngine` is the only supported engine right
now anyway…).