Reusable memory map pool
For the memory maps themselves, for the time being, we’re using a scheme
that uses `SHARED` maps which gives us a large chunk of the speed
beenfit described in
https://kazlauskas.me/entries/fast-page-maps-for-jit. Ideally we would
use a memfd based approach, however that turns out to require some
significant reworking of how the compiled functions are referenced – for
exapmle we want to apply linking and relocation on the writable view of
the functions, but the function references will contain references to
the executable mapping right after the `allocate` call.

I imagine we won’t get to implementing an improvement here for a little
while, and definitely not before we implement flat executable file
representations.